complaint delete function finalized v.2.1

+++ WebContent/readCompaint-demo.jsp
@@ -59,7 +59,7 @@ try{
 <td class="text-center"><%=resultSet.getString("complaint") %></td>
 <td class="text-center"><%=resultSet.getString("status") %></td>
 <td class="text-center"><a href="updateComplaint.jsp?cid=<%= resultSet.getInt("cid")%>" class="btn btn-warning btn-sm">Update</a></td>
-<td class="text-center"><a href="deleteComplaint?cid=<%= resultSet.getInt("cid")%>" class="btn btn-danger btn-sm">Delete</a></td>
+<td class="text-center"><a href="deleteComplaint?cid=<%= resultSet.getInt("cid")%>" class="btn btn-danger btn-sm" onclick="return confirm('Are you sure you want to delete this complaint?');">Delete</a></td>
 
 
 </tr>
